feat: 금액에 대한 정보를 BigDecimal로 변경

+++ src/main/java/cart/application/dto/order/OrderRefundResponse.java
@@ -1,14 +1,16 @@
 package cart.application.dto.order;
 
+import java.math.BigDecimal;
+
 public class OrderRefundResponse {
 
-    private final Integer refundPrice;
+    private final BigDecimal refundPrice;
 
-    public OrderRefundResponse(final Integer refundPrice) {
+    public OrderRefundResponse(final BigDecimal refundPrice) {
         this.refundPrice = refundPrice;
     }
 
-    public Integer getRefundPrice() {
+    public BigDecimal getRefundPrice() {
         return refundPrice;
     }
 }
