+++ src/main/java/cart/domain/order/Order.java
@@ -3,6 +3,7 @@ package cart.domain.order;
 import cart.domain.cartitem.dto.CartItemWithId;
 import cart.domain.coupon.dto.CouponWithId;
 import cart.domain.member.dto.MemberWithId;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -19,9 +20,9 @@ public interface Order {
 
     LocalDateTime getOrderedAt();
 
-    Integer getTotalPrice();
+    BigDecimal getTotalPrice();
 
-    Integer getDiscountedTotalPrice();
+    BigDecimal getDiscountedTotalPrice();
 
     Boolean isValid();
 }
