+++ src/main/java/cart/domain/refund/RefundPolicy.java
@@ -1,11 +1,12 @@
 package cart.domain.refund;
 
 import cart.domain.order.Order;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface RefundPolicy {
 
     boolean isAvailable(final Order order, final LocalDateTime currentTime);
 
-    int calculatePrice(final int price);
+    BigDecimal calculatePrice(final BigDecimal price);
 }
